refactoring common log attributes retrieval
